reorganize into common folder
