+++ lib/data/prefs.dart
@@ -13,11 +13,11 @@
 // Cuppa preferences
 // - Handle shared prefs
 
-import 'package:cuppa_mobile/data/constants.dart';
-import 'package:cuppa_mobile/data/globals.dart';
+import 'package:cuppa_mobile/common/constants.dart';
+import 'package:cuppa_mobile/common/globals.dart';
+import 'package:cuppa_mobile/common/helpers.dart';
 import 'package:cuppa_mobile/data/localization.dart';
 import 'package:cuppa_mobile/data/tea.dart';
-import 'package:cuppa_mobile/helpers.dart';
 
 import 'dart:convert';
 import 'package:flutter/material.dart';
