+++ lib/data/tea_timer.dart
@@ -13,7 +13,7 @@
 // Cuppa data
 // - Instance of a tea timer
 
-import 'package:cuppa_mobile/helpers.dart';
+import 'package:cuppa_mobile/common/helpers.dart';
 import 'package:cuppa_mobile/data/tea.dart';
 
 import 'dart:async';
