+++ lib/main.dart
@@ -12,12 +12,12 @@
 
 // Cuppa: a simple tea timer app for Android and iOS
 
-import 'package:cuppa_mobile/data/constants.dart';
-import 'package:cuppa_mobile/data/globals.dart';
+import 'package:cuppa_mobile/common/constants.dart';
+import 'package:cuppa_mobile/common/globals.dart';
+import 'package:cuppa_mobile/common/themes.dart';
 import 'package:cuppa_mobile/data/localization.dart';
 import 'package:cuppa_mobile/data/prefs.dart';
 import 'package:cuppa_mobile/data/provider.dart';
-import 'package:cuppa_mobile/widgets/themes.dart';
 import 'package:cuppa_mobile/widgets/timer_page.dart';
 
 import 'package:dynamic_color/dynamic_color.dart';
