+++ lib/widgets/about_page.dart
@@ -14,14 +14,15 @@
 // - Version and build number
 // - Links to GitHub, Weblate, etc.
 
-import 'package:cuppa_mobile/data/constants.dart';
-import 'package:cuppa_mobile/data/globals.dart';
+import 'package:cuppa_mobile/common/constants.dart';
+import 'package:cuppa_mobile/common/globals.dart';
+import 'package:cuppa_mobile/common/icons.dart';
+import 'package:cuppa_mobile/common/text_styles.dart';
 import 'package:cuppa_mobile/data/localization.dart';
 import 'package:cuppa_mobile/data/provider.dart';
-import 'package:cuppa_mobile/widgets/common.dart';
+import 'package:cuppa_mobile/widgets/list_divider.dart';
 import 'package:cuppa_mobile/widgets/platform_adaptive.dart';
 import 'package:cuppa_mobile/widgets/stats_page.dart';
-import 'package:cuppa_mobile/widgets/text_styles.dart';
 import 'package:cuppa_mobile/widgets/tutorial.dart';
 
 import 'package:flutter/material.dart';
@@ -150,7 +151,7 @@ class AboutWidget extends StatelessWidget {
                 child: Container(
                   margin: const EdgeInsets.only(top: 12.0),
                   // About text linking to app website
-                  child: aboutText(),
+                  child: _aboutText(),
                 ),
               ),
             ),
@@ -181,6 +182,31 @@ class AboutWidget extends StatelessWidget {
             const EdgeInsets.symmetric(vertical: 6.0, horizontal: 12.0),
         dense: true,
       ),
+    );
+  }
+
+  // About text linking to app website
+  Widget _aboutText() {
+    return InkWell(
+      child: Container(
+        padding: const EdgeInsets.symmetric(vertical: 6.0, horizontal: 12.0),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            Text(AppString.about_app.translate(), style: textStyleFooter),
+            const Row(
+              children: [
+                Text(aboutCopyright, style: textStyleFooter),
+                VerticalDivider(),
+                Text(aboutURL, style: textStyleFooterLink),
+              ],
+            ),
+          ],
+        ),
+      ),
+      onTap: () =>
+          launchUrl(Uri.parse(aboutURL), mode: LaunchMode.externalApplication),
     );
   }
 }
