+++ lib/widgets/list_divider.dart
@@ -1,0 +1,22 @@
+/*
+ *******************************************************************************
+ Package:  cuppa_mobile
+ Class:    list_divider.dart
+ Author:   Nathan Cosgray | https://www.nathanatos.com
+ -------------------------------------------------------------------------------
+ Copyright (c) 2017-2023 Nathan Cosgray. All rights reserved.
+
+ This source code is licensed under the BSD-style license found in LICENSE.txt.
+ *******************************************************************************
+*/
+
+// Cuppa list divider widget
+
+import 'package:flutter/material.dart';
+
+// List divider
+const Widget listDivider = Divider(
+  thickness: 1.0,
+  indent: 12.0,
+  endIndent: 12.0,
+);
