+++ lib/widgets/tea_color_dialog.dart
@@ -13,7 +13,8 @@
 // Cuppa tea color selection dialog
 
 import 'package:cuppa_mobile/data/tea.dart';
-import 'package:cuppa_mobile/widgets/common.dart';
+import 'package:cuppa_mobile/widgets/list_divider.dart';
+import 'package:cuppa_mobile/widgets/mini_tea_button.dart';
 import 'package:cuppa_mobile/widgets/platform_adaptive.dart';
 
 import 'package:flex_color_picker/flex_color_picker.dart';
